Clean up GUID serialization to avoid boxing
