+++ test/Serde.Generation.Test/test_output/AllInOneTest.GeneratorTest/Serde.Test.AllInOne.ISerialize.g.verified.cs
@@ -30,11 +30,11 @@ partial record AllInOne
             _l_type.WriteI64(_l_info, 10, value.LongField);
             _l_type.WriteI128(_l_info, 11, value.Int128Field);
             _l_type.WriteString(_l_info, 12, value.StringField);
-            _l_type.WriteBoxedValue<global::System.DateTimeOffset, global::Serde.DateTimeOffsetProxy>(_l_info, 13, value.DateTimeOffsetField);
+            _l_type.WriteDateTimeOffset(_l_info, 13, value.DateTimeOffsetField);
             _l_type.WriteDateTime(_l_info, 14, value.DateTimeField);
-            _l_type.WriteBoxedValue<global::System.DateOnly, global::Serde.DateOnlyProxy>(_l_info, 15, value.DateOnlyField);
-            _l_type.WriteBoxedValue<global::System.TimeOnly, global::Serde.TimeOnlyProxy>(_l_info, 16, value.TimeOnlyField);
-            _l_type.WriteBoxedValue<global::System.Guid, global::Serde.GuidProxy>(_l_info, 17, value.GuidField);
+            _l_type.WriteDateOnly(_l_info, 15, value.DateOnlyField);
+            _l_type.WriteTimeOnly(_l_info, 16, value.TimeOnlyField);
+            _l_type.WriteGuid(_l_info, 17, value.GuidField);
             _l_type.WriteString(_l_info, 18, value.EscapedStringField);
             _l_type.WriteStringIfNotNull(_l_info, 19, value.NullStringField);
             _l_type.WriteValue<uint[], Serde.ArrayProxy.Ser<uint, global::Serde.U32Proxy>>(_l_info, 20, value.UIntArr);
